requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
arclet-alconna==1.8.34
arclet-alconna-tools==0.7.10
arrow==1.3.0
asyncio==3.4.3
asyncio-dgram==2.2.0
attrs==24.2.0
bbcode==1.1.0
bilireq==0.2.13
binaryornot==0.4.4
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
cashews==7.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cookiecutter==2.6.0
cookit==0.8.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
dateparser==1.2.0
ddddocr==1.5.6
distlib==0.3.9
dnspython==2.7.0
dulwich==0.21.7
emoji==2.14.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fleep==1.0.1
fluidsynth==0.2
fonttools==4.55.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.4
keyring==24.3.1
kiwisolver==1.4.7
loguru==0.7.2
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mcstatus==11.1.1
mdurl==0.1.2
meme-generator==0.1.6
midi2audio==0.1.1
MIDIUtil==1.2.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nb-cli==1.4.2
nepattern==0.7.7
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.54.0
nonebot-plugin-apscheduler==0.4.0
nonebot-plugin-callapi==0.2.1
nonebot-plugin-capoo==0.1.5
nonebot-plugin-datastore==1.3.0
nonebot-plugin-gscode==0.2.2
nonebot-plugin-gspanel==0.2.25
nonebot-plugin-heweather==0.8.2
nonebot-plugin-htmlrender==0.3.5
nonebot-plugin-localstore==0.7.2
nonebot-plugin-miragetank==0.2.1
nonebot-plugin-multincm==1.1.3
nonebot-plugin-orm==0.7.6
nonebot-plugin-picstatus==2.0.0.post4
nonebot-plugin-send-anything-anywhere==0.7.1
nonebot-plugin-waiter==0.8.0
nonebot2==2.4.0
nonebot_plugin_alias==0.4.0
nonebot_plugin_code==0.0.8
nonebot_plugin_emojimix==0.4.0
nonebot_plugin_imagetools==0.4.2
nonebot_plugin_makemidi==0.3.0
nonebot_plugin_memes==0.7.11
nonebot_plugin_session==0.3.2
nonebot_plugin_session_orm==0.2.0
nonebot_plugin_setu4==0.15.114514.2
nonebot_plugin_userinfo==0.2.6
noneprompt==0.1.9
numpy==1.26.4
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pil-utils==0.1.10
pillow==10.4.0
pkginfo==1.11.2
platformdirs==4.3.6
playwright==1.49.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
pyee==12.0.0
pyfiglet==1.0.2
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.10.0
pymdown-extensions==10.12
PyMySQL==1.1.1
pyncm==1.7.1
pyparsing==3.2.0
pypinyin==0.53.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-markdown-math==0.8
python-multipart==0.0.17
python-slugify==8.0.4
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
revDeepSeek==0.0.1
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
silk-python==0.2.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
StrEnum==0.4.15
sympy==1.13.3
tarina==0.6.7
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.2
trio==0.27.0
trio-websocket==0.11.1
trove-classifiers==2024.10.21.16
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.27.1
wasmtime==29.0.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
win32-setctime==1.1.0
wolframalpha==5.1.3
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.0
zipp==3.21.0
